requirements.txt
Flask
rtree
pyarrow
dask[dataframe]
pandas
mysql-connector-python
numpy
sqlalchemy
mysqlclient
folium